% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_modelFUNCTIONS.R
\name{conres}
\alias{conres}
\title{Dynamic function for input into ode solver}
\usage{
conres(t, states, par)
}
\arguments{
\item{t}{time sequence for which output is wanted}

\item{states}{initial biomass values for the species in the system}

\item{par}{list of named parameters for input into the model}
}

