% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_modelFUNCTIONS.R
\name{conversion}
\alias{conversion}
\title{Converts directed matrix to undirected}
\usage{
conversion(tm)
}
\arguments{
\item{tm}{Food web adjacency matrix}
}
\value{
A new adjacency matrix where if a_ij  = 1 so does a_ji
}

