% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_modelFUNCTIONS.R
\name{eventfun}
\alias{eventfun}
\title{Extinction event function for the ODE system}
\usage{
eventfun(times, states, parms)
}
\arguments{
\item{times}{time sequence for which output is wanted}

\item{states}{biomass values for the species in the system}

\item{parms}{additional required parameters for the event}
}

