% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_modelFUNCTIONS.R
\name{Fbd}
\alias{Fbd}
\title{Consumer-Interference functional response}
\usage{
Fbd(B, A, B.0, xpar)
}
\arguments{
\item{B}{vector of biomasses}

\item{A}{binary adjacency matrix of the food web}

\item{B.0}{half saturation constant}

\item{xpar}{control parameter that alters the strength of interference}
}

