% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_modelFUNCTIONS.R
\name{G.i}
\alias{G.i}
\title{Basal growth function}
\usage{
G.i(r, B, K)
}
\arguments{
\item{r}{Growth rate in the absence of predators}

\item{B}{Biomass}

\item{K}{Carrying capacity}
}
\value{
numeric vector of population sizes
}
\description{
\code{G.i} returns population growth of basal species in the absence of predation
}

