% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_modelFUNCTIONS.R
\name{get.r}
\alias{get.r}
\title{Create vector of non-zero growth rates for basal species}
\usage{
get.r(amat)
}
\arguments{
\item{amat}{Adjacency matrix of food web}
}
\value{
numeric vector of 1 if species is basal, and 0 otherwise
}

