% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_modelFUNCTIONS.R
\name{motifCounter3}
\alias{motifCounter3}
\title{Find the counts of three species configurations through time from \code{CRsimulator}}
\usage{
motifCounter3(dyn, web)
}
\arguments{
\item{dyn}{Matrix of biomass dynamics from \code{CRsimulator}}

\item{web}{Initial food web adjacency matrix}
}
\value{
A dataframe of counts of three species configurations where each row is a time step in the model
}

