% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_modelFUNCTIONS.R
\name{trophicChange}
\alias{trophicChange}
\title{Change in trophic level through time}
\usage{
trophicChange(dyn, web)
}
\arguments{
\item{dyn}{Matrix of biomass dynamics from \code{CRsimulator}}

\item{web}{Initial food web adjacency matrix}
}
\value{
Matrix of trophic position for each species in the food web
}
\description{
Change in trophic level through time
}

