% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_modelFUNCTIONS.R
\name{WEBind}
\alias{WEBind}
\title{Get common food web structural indices}
\usage{
WEBind(dyn, web)
}
\arguments{
\item{dyn}{Matrix of biomass dynamics from \code{CRsimulator}}

\item{web}{Initial food web adjacency matrix}
}
\value{
A matrix of food web indices where each row is a time step in the dynamics simulation
}

